Editieren des Profilbild in Settings.
