+++ CoolGag/imports/ui/components/profile/ProfileHeader.js
@@ -51,21 +51,15 @@ class ProfileHeader extends React.Component {
 				<h1>{this.props.data.user.name}</h1>
 
 				{this.props.data.user.profilePic && this.props.data.user.profilePic.url && 
-					<div className="defaultImage">
-						<span>
-							<span>
-								<input className="upload pa3 bn ttu pointer bg-black-10 dim" type="button" value="Update" onClick={this.startChoosingImage.bind(this)} />
-								<input className="delete pa3 bn ttu pointer bg-black-10 dim" type="button" value="Delete" onClick={this.handleDelete.bind(this)} />
-							</span>
-							<img src={this.props.data.user.profilePic.url} crossOrigin='Anonymous' role='presentation' className='w-100 profilePic' onError={this.onProfileImageLoadError.bind(this)} />
-						</span>	
-					</div>	
+					<div className="profileImage">
+						<img src={this.props.data.user.profilePic.url} crossOrigin='Anonymous' role='presentation' className='w-100 profilePic' onError={this.onProfileImageLoadError.bind(this)} />
+					</div>
 				}
 				{!(this.props.data.user.profilePic && this.props.data.user.profilePic.url) &&
 					<div className="defaultImage">
 						<span>
-							<div className="input-group">
-								<input className="upload pa3 bn ttu pointer bg-black-10 dim" type="button" value="Upload a picture" onClick={this.startChoosingImage.bind(this)} />
+							<div className="ínput-center">
+								<input className="pa3 bn ttu pointer bg-black-10 dim" type="button" value="Upload a profile picture" onClick={this.changeImage.bind(this)} />
 							</div>
 							<svg width="50%" height="50%" viewBox="0 0 100 100" version="1.1" xmlns="http://www.w3.org/2000/svg">
 								<path d="M65.904,52.834c-4.734,3.725-10.695,5.955-17.172,5.955c-6.316,0-12.146-2.119-16.821-5.68C16.654,55.575,5,68.803,5,84.757 c0,17.711,14.356,6.197,32.065,6.197h25.868C80.643,90.954,95,102.468,95,84.757C95,68.051,82.22,54.333,65.904,52.834z" fill="#5a0000"/>
@@ -78,58 +72,6 @@ class ProfileHeader extends React.Component {
 					</div>		
 				}
 
-				<form style={{ maxWidth: 400 }} className='' onSubmit={this.handleUpload.bind(this)}>
-					{this.state.isEditingPicture &&
-						<div>
-							<FileSelectButton onSelect={this.handleFileSelect.bind(this)} />
-							<WindowDropZone
-								onDragStart={this.onDragStart.bind(this)}
-								onDragEnd={this.onDragEnd.bind(this)}
-								onDrop={this.onDropFiles.bind(this)}
-							/>
-
-							{ !this.state.imageUrl &&
-								<div className='w-100 dropzone mv3'>
-									{ !this.state.isLoadingFile && !this.state.isDraggingFile &&
-										<span>Select a file or drop an image here.</span>
-									}
-									{ this.state.isLoadingFile &&
-										<span>Processing File...</span>
-									}
-									{ this.state.isDraggingFile && this.state.isValidType &&
-										<span>Drop to Upload</span>
-									}
-									{ this.state.isDraggingFile && !this.state.isValidType &&
-										<span>Invalid File</span>
-									}
-								</div>
-							}
-							{ this.state.imageUrl &&		
-								<div className={'imagePreviewCotnainer w-100 mv3' + (this.state.isDraggingFile ? ' isDragging' : '')}>
-									<div className={'imagePreview' + (this.state.isTextEntered ? ' textEntered' : '')}>
-										<img src={this.state.imageUrl} crossOrigin='Anonymous' role='presentation' className='w-100' onLoad={this.onImageLoaded.bind(this)} onError={this.onImageLoadError.bind(this)} />
-									</div>
-									{ (this.state.isDraggingFile || this.state.isLoadingFile) &&
-										<div className='w-100 dropzone'>
-											{ this.state.isDraggingFile && this.state.isValidType &&
-												<span>Drop to Upload</span>
-											}
-											{ this.state.isDraggingFile && !this.state.isValidType &&
-												<span>Invalid File</span>
-											}
-											{ this.state.isLoadingFile &&
-												<span>Processing File...</span>
-											}
-										</div>
-									}
-								</div>
-							}
-							<button type="submit" disabled={(this.isSubmittable() ? "" : "disabled")} className={'pa3 bn ttu pointer' + (this.isSubmittable() ? " bg-black-10 dim" : " black-30 bg-black-05 disabled")}>
-								{this.state.isSubmitting ? (this.state.isRendering ? 'Rendering...' : 'Submitting ...') : 'Okay'}
-							</button>
-						</div>
-					}
-				</form>
 				<div>
 					Member since {this.state.createdAt}.
 				</div>
@@ -143,127 +85,14 @@ class ProfileHeader extends React.Component {
 		);
 	}
 
-	startChoosingImage() {
-		this.setState({
-			isEditingPicture: true
-		});
-	}
-
 	onProfileImageLoadError() {
-		this.props.data.user.profilePic = null;
+	  this.props.data.user.profilePic = null;
 	}
 
-	isSubmittable() {
-		return this.state.file && this.state.imageUrl && !this.state.isSubmitting;
+	changeImage() {
+		this.props.router.replace('/settings');
 	}
 
-
-	onDragStart(validType) {
-		this.setState({
-			isDraggingFile: true,
-			isValidType: validType
-		});
-	}
-	onDragEnd() {
-		this.setState({
-			isDraggingFile: false
-		});
-	}
-	onDropFiles(files) {
-		if(files.length >= 1) {
-			this.handleFileSelect(files[0]);
-		}
-	}
-
-	onImageLoadError(event) {
-		console.log('error');
-		this.setState({'imageUrl': ''});
-		this.setState({'file': null});
-	}
-
-	onImageLoaded(event) {
-		var imageElement = event.nativeEvent.srcElement || event.nativeEvent.originalTarget;
-		$('.uncheckedSpelling').attr('spellcheck', 'false');
-		this.setState({
-			imageSize: {width: imageElement.naturalWidth, height: imageElement.naturalHeight},
-			isLoadingFile: false
-		});
-	}
-	
-	handleFileSelect(file) {
-		if(file != null) {
-			this.setState({
-				file: null,
-				imageUrl: '',
-				isLoadingFile: true,
-			});
-			FileHandling.getDataUrl(file, (result) => {
-				this.setState({
-					file: file,
-					imageUrl: result,
-					isLoadingFile: false
-				});
-			});
-		}
-	}
-
-	handleUpload(event) {
-		event.preventDefault();
-		this.setState({'isSubmitting': true});
-
-		var continueUpload = () => {
-			fetch('https://api.graph.cool/file/v1/cj2ryvxmbt4qw0160y6qhdgdl', {
-				body: data,
-				method: 'POST'
-			}).then((response) => {
-				response.json().then(result => {
-					//this.setState({imageUrl: result.url});
-					this.setState({postedFileId: result.id});
-					this.setState({userId: this.props.data.user.id});
-					var {postedFileId, userId} = this.state;
-					this.props.changeProfilePic({
-						variables: {
-							postedFileId: postedFileId,
-							userId: userId
-						}
-					});
-					this.setState({'isSubmitting': false});
-				});
-			}).catch((exception) => {
-				// TODO: handle upload error
-				console.log('error uploading the profile picture!');
-				this.setState({'isSubmitting': false});
-			});
-		};
-		
-		let data = new FormData();
-		data.append('data', this.state.file);
-		continueUpload();
-
-		this.setState({'isEditingPicture': false});
-		return false;
-	}
-
-	handleDelete(event) {
-		event.preventDefault();
-		this.setState({'isSubmitting': true});
-
-		this.setState({fileId: this.props.data.user.profilePic.id});
-		this.setState({userId: this.props.data.user.id});
-		var {fileId, userId} = this.state;
-		this.props.deleteProfilePic({
-			variables: {
-				userId: userId,
-				fileId: fileId
-			}
-		}).catch((exception) => {
-			// TODO: handle upload error
-			console.log('error deleting the profile picture!');
-			this.setState({'isSubmitting': false});
-		});
-		this.setState({'isSubmitting': false, 'isEditingPicture': false});
-		return false;
-	}
 }
 
 const profileData = gql`
@@ -280,43 +109,5 @@ const profileData = gql`
 		}
 	}
 `
-
-const uploadPicture = gql`
-	mutation ($userId: ID!, $postedFileId: ID!) {
-		setUserOnFile(
-			userProfilePicUserId: $userId,
-			profilePicFileId: $postedFileId
-		) 
-		{
-			userProfilePicUser {
-				id
-			}
-		}
-	}
-`
-
-const deletePicture = gql`
-	mutation ($userId: ID!, $fileId: ID!) {
-		unsetUserOnFile(
-			userProfilePicUserId: $userId,
-			profilePicFileId: $fileId
-		) 
-		{
-			userProfilePicUser {
-				id
-			}
-		}
-		deleteFile(
-			id: $fileId
-		)
-		{
-			id
-		}
-	}
-`
-
-export default compose(
-  graphql(uploadPicture, { name: 'changeProfilePic' } ),
-  graphql(deletePicture, { name: 'deleteProfilePic'} ),
-  graphql(profileData, { options: { forceFetch: true }} )
-)(withApollo(withRouter(ProfileHeader)))
+export default graphql(profileData, { options: { forceFetch: true }} )
+(withApollo(withRouter(ProfileHeader)))
