+++ CoolGag/imports/ui/containers/profileLists/MyGroups.js
@@ -1,14 +1,19 @@
 import gql from 'graphql-tag';
 import { graphql } from 'react-apollo';
-import ListPage from '../../components/ListPage';
+import GroupListPage from '../../components/profile/GroupListPage';
 
 //TODO
 const MyGroupsQuery = gql`query {
-  allPosts(orderBy: createdAt_DESC) {
-    id
-	postedFile { url }
-    description
-  }
+   user{
+       groups{
+           id
+            name
+           users{
+               id
+               name
+           }
+       }
+   }
 }`
 
-export default graphql(MyGroupsQuery)(ListPage);
+export default graphql(MyGroupsQuery)(GroupListPage);
